Quotes parsing API changed from AlphaVantage to Pandas-Datareader. AlphaVantage is now used to get quotes symbols list.

+++ quotes/models.py
@@ -79,7 +79,7 @@ class Quotes(models.Model):
                     'high': float(data[key]['2. high']),
                     'low': float(data[key]['3. low']),
                     'close': float(data[key]['4. close']),
-                    'volume': float(data[key]['5. volume']),
+                    'volume': int(data[key]['5. volume']),
                     'non-trading': False
                 }
             else:
